implement simple aggregation scheme
This aggregation scheme is intended to allow OpenMPI to transfer larger messages
if the user-reported partitions are too small or too many.
This is achieved by using an internal partitioning where each internal (transfer)
partition corresponds to one or multiple user-reported partitions.

The implementation provides a queue-like interface, where user-partitions can be
inserted and transfer partitions can be extracted as soon as all corresponding
user-partitions are ready.

This is achieved by associating each transfer partition with an atomic counter,
tracking the number of corresponding pready-calls. As soon as a counter reaches
the number of corresponding user-partitions, the transfer partition is inserted
into a ring buffer, from where it can be extracted.

This implementation is thread-safe.