use aggregation scheme in part_persist_aggregated component
A few changes are made to the component to allow for using a different
partitioning internally than reported by the user:

During initialization, an internal partitioning is computed from the
user-provided one, taking into account the limits imposed by
the mca-parameters on partition size and count.
This internal partitioning can be equal to or less fine-grained
than the one provided by the user.

The sender side request objects now need to hold state for message aggregation,
therefore an aggregation_state field is added. This is not required for receive
side requests, as the receiver side code already supports differing send-side
partitionings.

As transfer partition sizes might not divide user-partition sizes, the last
transfer partition can correspond to fewer user-partitions than the other ones.
For this reason, a field (real_remainder) is introduced into request objects that
holds the number of elements corresponding to the last transfer
partition.
This field is also added to the setup information exchanged between sender and
receiver during initialization.

To account for this potentially smaller last
transfer partition, the initialization of the internal persistent requests is
adjusted.

Signed-off-by: Axel Schneewind <axel.schneewind@hlrs.de>